refactor: random string utils 를 공통 모듈로 이동
