+++ bob-works-authorization-service/src/main/java/org/bobpark/authorizationservice/domain/authorization/entity/AuthorizationClient.java
@@ -1,6 +1,7 @@
 package org.bobpark.authorizationservice.domain.authorization.entity;
 
 import static org.apache.commons.lang3.ObjectUtils.*;
+import static org.bobpark.core.utils.RandomUtils.*;
 
 import java.time.LocalDateTime;
 import java.util.ArrayList;
@@ -24,7 +25,6 @@ import lombok.ToString;
 import lombok.ToString.Exclude;
 
 import org.bobpark.authorizationservice.common.entity.BaseTimeEntity;
-import org.bobpark.authorizationservice.common.utils.random.RandomUtils;
 
 @ToString
 @Getter
@@ -59,7 +59,7 @@ public class AuthorizationClient extends BaseTimeEntity {
         Boolean requiredAuthorizationConsent, Long accessTokenTimeToLive) {
         this.id = id;
         this.clientId = defaultIfNull(clientId, UUID.randomUUID().toString());
-        this.clientSecret = String.format("{noop}%s", defaultIfNull(clientSecret, RandomUtils.randomString(20)));
+        this.clientSecret = String.format("{noop}%s", defaultIfNull(clientSecret, randomString(20)));
         this.clientName = clientName;
         this.clientIssueAt = defaultIfNull(clientIssueAt, LocalDateTime.now());
         this.clientSecretExpiresAt = clientSecretExpiresAt;
