+++ bob-works-core/src/main/java/org/bobpark/core/utils/RandomUnit.java
@@ -1,4 +1,4 @@
-package org.bobpark.authorizationservice.common.utils.random;
+package org.bobpark.core.utils;
 
 import java.util.Arrays;
