+++ bob-works-core/src/main/java/org/bobpark/core/utils/RandomUtils.java
@@ -1,4 +1,4 @@
-package org.bobpark.authorizationservice.common.utils.random;
+package org.bobpark.core.utils;
 
 import java.util.Random;
